+++ scripts/run_Eager.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/env bash
 
+## DEPENDENCY
+pandora_helper="/mnt/archgen/tools/helper_scripts/py_helpers/pyPandoraHelper/pyPandoraHelper.py"
+
 ## Defaults
 rush=''
 array=''
@@ -42,20 +45,10 @@ for analysis_type in "SG" "TF" "RP" "RM"; do
     for eager_input in ${root_input_dir}/${analysis_type}/*/*/*.tsv; do
         ## Set output directory name from eager input name
         ind_id=$(basename ${eager_input} .tsv)
-        site_id="${ind_id:0:3}"
+        site_id=`${pandora_helper} -g site_id ${ind_id}` ## Site inferred by pyPandoraHelper
         eager_output_dir="${root_output_dir}/${analysis_type}/${site_id}/${ind_id}"
 
         run_name="-resume" ## To be changed once/if a way to give informative run names becomes available
-        
-        ## TODO Give informative run names for easier trackingin tower.nf
-        ##  If the output directory exists, assume you need to resume a run, else just name it
-        # if [[ -d "${eager_output_dir}" ]]; then
-        #     command_string="-resume"
-        # else
-        #     command_string="-name"
-        # fi
-        # ## Run name is individual ID followed by analysis_type. -resume or -name added as appropriate
-        # run_name="${command_string} $(basename ${eager_input} .tsv)_${analysis_type}"
 
         ## If no multiqc_report exists (last step of eager), or TSV is newer than the report, start an eager run.
         #### Always running with resume will ensure runs are only ever resumed instead of restarting.